29/06/2021 23:45

+++ nodejs-complete-guide/controllers/admin.js
@@ -1,5 +1,6 @@
 const { validationResult } = require('express-validator/check');
 const mongoose = require('mongoose');
+const fileHelper = require('../util/file');
 
 const Product = require('../models/product');
 exports.getAddProduct = (req, res, next) => {
@@ -19,9 +20,24 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
+
+    if (!image) {
+        return res.status(422).render('admin/edit-product', {
+            pageTitle: 'Add Product',
+            path: '/admin/add-product',
+            editing: false,
+            hasError: true,
+            errorMessage: 'Attached file is not an image.',
+            product: {
+                ...req.body,
+            },
+            validationErrors: [],
+        });
+    }
+
     const errors = validationResult(req);
 
     if (!errors.isEmpty()) {
@@ -38,8 +54,9 @@ exports.postAddProduct = (req, res, next) => {
         });
     }
 
+    const imageUrl = image.path;
+
     const product = new Product({
-        _id: new mongoose.Types.ObjectId('60d0a9939266206414cb22f7'),
         title: title,
         price: price,
         description: description,
@@ -94,7 +111,7 @@ exports.getEditProduct = (req, res, next) => {
 exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId;
     const updatedTitle = req.body.title;
-    const updatedImageUrl = req.body.imageUrl;
+    const image = req.file;
     const updatedPrice = req.body.price;
     const updatedDescription = req.body.description;
 
@@ -125,7 +142,14 @@ exports.postEditProduct = (req, res, next) => {
             product.title = updatedTitle;
             product.price = updatedPrice;
             product.description = updatedDescription;
-            product.imageUrl = updatedImageUrl;
+            if (image) { // if a new image is updated
+                // delete the old image
+                // fire and forget
+                fileHelper.deleteFile(product.imageUrl);
+
+                // set the new imageUrl
+                product.imageUrl = image.path;
+            }
             return product.save()
                 .then(result => {
                     console.log('UPDATED PRODUCT!');
@@ -160,15 +184,24 @@ exports.getProducts = (req, res, next) => {
 };
 
 exports.postDeleteProduct = (req, res, next) => {
-    const prodId = req.body.productId;
-    Product.deleteOne({ _id: prodId, userId: req.user._id })
+    const prodId = req.params.productId;
+    Product.findById(prodId)
+        .then(product => {
+            if (!product) {
+                return next(new Error('Product not found.'));
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({ _id: prodId, userId: req.user._id });
+        })
         .then(() => {
-            console.log('DESTROY PRODUCT');
-            return res.redirect('/admin/products');
+            console.log('DESTROYED PRODUCT');
+            res.status(200).json({
+                message: "Success!",
+            });
         })
         .catch(err => {
-            const error = new Error(err);
-            error.httpStatus = 500;
-            return next(error);
+            res.status(500).json({
+                message: "Deleting product failed.",
+            });
         });
 };
